+++ demo/guihello/flyout.h
@@ -1,0 +1,5 @@
+/* Flyout window */
+
+#include <gui/gui.hxx>
+
+Panel *flyout_window(Window *window);